Remove as many multi-line comments as possible
I'm trying out lua-format and it treats multi-line comments as *inline*
comments (which I guess they can also be used as?) and it's considerably
easier to just remove all the stupid multi-line comments scattered
throughout this codebase then it is to attempt to get a tool changed.

In a couple of situations I just added -- at the end of multi-line
comments that I couldn't be bothered to deal with as that fairly hackily
turns them into a block and prevents mangling.

See-also: https://github.com/Koihik/LuaFormatter/issues/32