Add an autocomplete helper for vscode
This solves two problems
- The linter continuously bitching at me for unknown globals
- Intellisense not knowing all the available libs

This doesn't solve the intellisense problem particularly well since it
has absolutely no info about the functions, but it does at least now
know about the functions which is a start.

I also would quite like this not to be in the root of the gamemode but
the extension doesn't seem to be very happy about libraries that are not
in the workspace root.